Updated a lot of messages + fix reroll deletes shop

+++ src/decorators/delete.request.message.decorator.ts
@@ -6,12 +6,13 @@ const DeleteRequestMessage = () => (_target: object, _propKey: string, desc: Pro
 
   // eslint-disable-next-line no-param-reassign
   desc.value = async function wrapped(...args: ContextMessageUpdate[]): Promise<void> {
+    const isActionCommand = args[1];
     const botId = (await Bot.getInstance().app.telegram.getMe()).id;
     const botUser = await Bot.getInstance().app.telegram.getChatMember(args[0].chat!.id, botId);
 
     await method.apply(this, args);
 
-    if (botUser.can_delete_messages) {
+    if (botUser.can_delete_messages && !isActionCommand) {
       await args[0].deleteMessage();
     }
   };
